+++ gluecodium/src/test/resources/stubs_smoke/classes/output/android/com/example/smoke/SomeClass.java
@@ -7,19 +7,21 @@ public class SomeClass extends NativeBase {
     public SomeClass() {
     }
     public void voidFunction() {
+        throw RuntimeException("This method is a stub and has to be mocked.");
     }
     public boolean boolFunction() {
-        return false;
+        throw RuntimeException("This method is a stub and has to be mocked.");
     }
     public int intFunction() {
-        return 0;
+        throw RuntimeException("This method is a stub and has to be mocked.");
     }
     public String stringFunction() {
-        return null;
+        throw RuntimeException("This method is a stub and has to be mocked.");
     }
     public SomeClass classFunction() {
-        return null;
+        throw RuntimeException("This method is a stub and has to be mocked.");
     }
     public static void staticFunction() {
+        throw RuntimeException("This method is a stub and has to be mocked.");
     }
 }